WA-282 WhitelistedDapps actions and reducers refactoring
