+++ app/actions/whitelistedDapps/deleteWhitelistedDapp.js
